checking for 2nd element to move
checking for 2nd element to move

+++ lib/java_buildpack/component/base_component.rb
@@ -156,7 +156,9 @@ module JavaBuildpack
 
                 FileUtils.mkdir_p target_directory.parent
                 FileUtils.mv Pathname.new(root).children.first, target_directory
-                FileUtils.mv Pathname.new(root).children[1], target_directory
+                if Pathname.new(root).children[1].exist?
+                  FileUtils.mv Pathname.new(root).children[1], target_directory
+                end
               end
             else
               FileUtils.mkdir_p target_directory
